Remove testutil.ParseCgroupPaths
